Refactorized code & adding codebase for sequential augmentation
